requirements.txt
# Illumio Monitoring Tool - Dependencies
# Install with: pip install -r requirements.txt

# Async HTTP client
aiohttp>=3.9.0

# Configuration and validation
pyyaml>=6.0
pydantic>=2.0.0

# Data processing
pandas>=2.0.0

# Excel export
xlsxwriter>=3.1.0
openpyxl>=3.1.0

# PDF generation
reportlab>=4.0.0

# Charts and visualization
matplotlib>=3.7.0

# Optional: For development and testing
# pytest>=7.0.0
# pytest-asyncio>=0.21.0
# black>=23.0.0
# flake8>=6.0.0
